Show a cursor on long press
Digits field doesn't allow "paste" option when there's no
text there, since EditText doesn't allow that when the cursor
is invisible while we intentionally hide the cursor.

One side effect of this change is that users will see the cursor
regardless of "paste" status if they long-click the field.

Also hide the cursor when it is at the end of the text.

Bug: 5394377
Change-Id: I046db65ac869ab8de860448d4dfe05199f48406b