SearchView on Join
We can have the small feature with some tweaks.

Right now ContctEntryListFragment is assuming that the first
directory in the fragment is the default directory, while on Join
screen the first one is "suggested contacts" directory. This change
modifies the implementation so that the fragment remove all directories
after the "default" directory intsead.

Also remove the code using MatrixCursor since we don't rely on it
anymore. This improves the performance of Join screen.

Change-Id: Ic232e09fb65bf72d19b069ce047e5a32dcb9f4ad