Normalize the phone number before contact look up.
The contact look-up request does not include a way to specify a
country which the number may correspond to. In the call log, we might
have calls that were placed or received while in one country and their
numbers will not match the contacts when the user is in a different
country.

To fix this issue, normalize the phone number using the country stored
in the call log together with the number itself and use that to look
up the contact.

Bug: 5252190
Change-Id: I46471fa2acffd541a2192f6948a613c13484c2d1