Only show options menu items when loaded.
Previously, we were showing the "Trash" option in the action bar before
the data was loaded, because we already knew this was a voicemail from
looking at the call log content, but we waited until the contact
information was loaded before showing the other options.

This lead to the set of options to change when the data is loaded, which
was confusing for the user which might already be interacting with the
action bar by that time.

Instead, this commit disables all options by default and enables the
needed ones only when the data has been fully loaded.

Bug: 5461913
Change-Id: I93c21c4afbfed62d0cb1b3649ff24278bfabf289