Show "no contacts" in frequent list on tablet
- We need to show "no contacts" under the
"frequently contacted" header in the frequent
fragment in the People app on the tablet.

Since that header was originally the 0th item
in the ListView, to show an empty view we would
either have to:

1) Add a second fake item in the ListView that shows
"No contacts". This requires either returning
view type "empty view" or view type "frequent"
for position = 1 based on the cursor count.
Checking the cursor in getItemViewType()
would make this more than a UI change.

2) Move the "Frequently contacted" header
outside the ListView and make the ContactTileAdapter
show the empty view state when there are 0
frequently contacted people (instead of the
existing implementation which fakes it by returning
1 when the cursor count is 0).

- This CL uses the second option so that we don't
need to add a cursor dependency in getItemViewType()
of ContactTileAdapter.

- Added a TODO because UI changes like this highlight
the fact that we should do some refactoring to
ContactTileAdapter and ContactTileListFragment
so that it's easier to add specific functionality to
the favorites, frequent, or group list.

Bug: 5485640
Change-Id: I7a069be3cf21ba3a7a857d636130a2e426d626f8