Reset views when changing contacts on tablet
- On tablet, we re-use the same fragments for
different contacts because we have a 2-pane view.
We need to reset the state to prevent showing
the wrong tab or old scroll state information from
a previous contact.

- Resetting the list adapter is getting rid of the
scroll state information and resetting the tab
carousel / view pager gets us back to the "about"
tab.

- Store the contact URI in the saved instance
state bundle so that on rotation, we still
recognize that it could be the same contact
and restore scroll position.

Bug: 5529595
Change-Id: Ibd874799f926b70430e47518a4b995394fbfe547