Fix flicker on group detail page on tablet
- Occassionally, after doing a sync, sometimes the group
detail page will flicker on the tablet.

- Remove destroyLoader() call right before restartLoader().
According to the documentation in LoaderManagerImpl,
destroying a loader destroys any data associated with it,
which seems like it could potentially cause all the photos
to disappear in the group member ListView.
(Otherwise it's not clear why the photos would all
disappear instead of showing the ListView empty state.
The group title / group membership size are always
on screen, and never blank out.)

- Haven't seen group detail page photos flicker anymore.
Selecting different groups in the group list still works
properly. Tested on phones and tablet.

Bug: 5513239
Change-Id: I39684b884b850715ded91449072cb8ff15e6d928