Fix test for selecting an account.
The AllIntentsActivity cannot use the classes from the application under
test. Therefore, we cannot use the AccountTypeManager and related
classes.

Since for the purpose of testing whether the People app can handle a
specific intent, we do not to replicate the entire set of functionality
in the test app as it is in the SelectAccountDialogFragment in the
People app itself. Instead, use a simplified version, which just queries
the account manager.

Bug: 5571116
Change-Id: I4b07889887139c5f46a973a6f92e7cf5458b17e7