Fix wrong contact being selected after editing (after a search)
- The problem is that after finding a contact via search
and editing that contact, we launch a VIEW intent to view
the contact. Then sometimes, that contact is not in the
current contact filter, so we may have to enter single contact
mode.

To avoid that, go back to the HC functionality of preserving
search mode. Now after editing, we return to the PeopleActivity
and preserve search mode if that's where we came from.
This was the relevant change from early ICS that is being
partially reverted:
Ibbaaccb03ac7961784dfa7b15c246b8d6f6489a4

The reason it was changed was to support viewing the contact
after editing a contact when the edit request came from a 3rd
party app. This still works with this change by adding
an intent extra in ContactEditorActivity.

- After the editing is done, only launch the VIEW intent
for the contact if we didn't come from the detail page for
that contact (on phone and tablet). We can just finish
the editor activity. This saves some time with setup / configuring
fragments because we don't have to resolve the VIEW intent
(i.e. On tablet, previously the VIEW intent launched the
ContactDetailActivity --> redirected to PeopleActivity -->
resolved intent to show the right contact.
Now we can just receive the onActivityResult and refresh
the page).

- Renamed ContactDetailActivity intent key to be consistent

Bug: 5542197
Change-Id: Ieaa9c147beeaa7c40f34fc1689858642b47fdbf9