Intercept single contact mode and reset to last saved preference
Single contact mode is hard to get out of unless the user
changes the "contacts to display" filter setting.

Thus, whenever the PeopleActivity onStop() is called and the user
and reenters through onStart() -  we reset the contact filter back
to the last saved filter according to SharedPreferences.

Since ContactListFilterController is a singleton, we have to
be careful in making sure the filter state is consistent with
the all fragment filter state because there can be multiple
PeopleActivity instances (i.e. another app launched People).

Bug: 5526485
Change-Id: I748ffc158a2de778dd9ecce4ab7f132af196b9a3