Don't use saved list state if the app hasn't been killed
On the tablet, clicking on a contact from the groups tab
will show the contact list + details page for that contact.
However, if a sync happens afterwards then the list will
wrongly scroll to some previous scroll position.

The fix is to not update the local variable mListState
when we saveInstanceState() on the contact list fragment.
This is because mListState is used in onLoadFinished() and
will scroll the list to an old scroll position even after
the correct contact has been selected in the list (from
an incoming intent request). We should only update mListState
when we are restoring state and reading from the save
instance state bundle.

Tested with rotation and killing the app on phone and tablet.

Bug: 5525474
Change-Id: Ic217f1151d2d7850946625192d6c8f3f03437bf1