Let ContactTileRow ignore accessibility event
When "explore by touch" is turned on and a user clicks empty area
in contact tile section, tts speaks a contact name next to the
empty area. This happens because ContactTileRow tries to obtain
some meaningful information from the invalid tile next to the
empty space. This change forces the row to ignore accesibility
event.

Bug: 5559634
Change-Id: I7a994c38411c9d551400cd79eb9edb80abb324b3