Store contact info using country iso as well.
Currently we are using just the number to store contact information that
we look-up. However, the look-up result depends also on the country in
which the call was placed or received.

This can lead to issues if the result is not the same depending on the
contact. This is particularly problematic in case of upgrades: the call
log will contain a number of entries without a country iso value since
the field was introduced after the Gingerbread release.

Use the country iso both when storing the contact in the in-memory
cache, as well as when updating the cache stored in the call log itself.

Bug: 5638376
Change-Id: I4c77556395147f05c2524d93cc1c800943395b6c