Remove verbose logging.
This condition used to be an error, but it actually now possible, in
fact quite frequent. Remove it from the log.

We used to trigger the update only when a new voicemail is received.
In that case, there should always be at least a voicemail to notify
about.

However, we currently do this also if we request do dismiss the the
notifications, e.g., when opening the call log. This log is a bit too
common now (basically every time the user opens the call log), so
downgrade it to a comment, since it is no longer an error condition.

Bug: 5657809
Change-Id: Ic35cc306ceca3fbe1d3fe83076e5120d810c354b