Correctly format numbers in RTL languages.
When phone numbers are formatted together with text in an RTL language,
the numbers will themselves formatted RTL by default: if they have any
formatting (e.g., +, -, (, ), or spaces), then these will not be
formatted correctly, as they should be shown LTR as a whole. This works
by itself if the number is the only thing shown in a text field (since
phone numbers are always made of characters that are entirely LTR), but
fails if the number is formatted within some translated text.

This commit forces the phone number to always be shown LTR.

Also, make it more explicit that the strict is meant to be used with a
number, since the description and xliff tag contradicted the name of the
string.

Bug: 5429886
Change-Id: Ifd1835ecfa89725ea596d02d9945712c4f351ddf