Avoid "Add New" button following the empty date field in new contact.
EventFieldEditorView.isEmpty() has two issues.  First, it doesn't deal
properly with the placeholder text in the empty date field.  Second,
it queries the text from the text field instead of from the edit-field
(i.e. from the view rather than from the model).  The latter leads to a
problem when isEmpty() is called before the text-field's value has been
set.  Both of these are now fixed.

Bug: 5265651
Change-Id: I8cf1a66ddfbeaca04211e227a20640556fb095b9