Mark intent test LIST_CONTACTS_WITH_PHONES_ACTION as deprecated
The action com.android.contacts.action.LIST_CONTACTS_WITH_PHONES
is hidden in ContactsContract. This action exists in the legacy
Contacts contract, but was marked as deprecated.

There's no functionality in DialtactsActivity to launch
search fragment immediately or to go to the favorites tab,
and even if we had it working on the phone devices, the intent
behavior would have to be handled differently on the tablet where
there is no phone app. Launching a picker isn't exactly the right
behavior because the calling app is probably not expecting a result.

Now this intent basically just shows the contact list, which is fine.
Just mark the intent test as deprecated, but in the future, we
should just make sure it doesn't crash.

Bug: 5150839
Change-Id: I26dfa941e59f3a33d9ec911df65287db389cc872