Prevent auto scroll of contact list after contact editing
Turn off smooth scroll to requested contact, but leave
the code in still.

After editing a contact on the tablet (and returning
a result to the PeopleActivity), don't reload the data
because that restarts the loader (which destroys
the existing loader and causes the scroll position in the
ListView to be lost).

Bug: 5739863
Change-Id: I32051494b9c6155ac05b0ab3e6b962d34e95bea5