Delay clearing digits until onStop()
It is a bit astonishing when users see digits being cleared
just after pressing "call" button on dialer screen, while
we're sure we should clear the previous digits entered when
a phone call is made.

Because onDestroy() may not be called even after the Activity's
being done, check the flag on onResume() too.

TESTED:
- make a phone call
- access voicemail service
- enter phone number and exit dialer without calling
- play with swipe (which may destroy the fragment)

Bug: 1745781
Change-Id: Ie0bc0ee8d5b1bb55bfea3b394b86cd31799c8238