Reduce call-log jankiness while scrolling.
Before this fix, we were triggering many extraneous relayouts.
The cause was that we populated the ContactInfo cache with empty
entries; when we compared against these to see if they had changed,
the answer was always "yes".  This happened even in bindView() when
scrolling new rows onscreen, with predictably janky results.

Bug: 6077964
Change-Id: I46b7795b38030b0d90f9d65c05c02dbad5776b37