Don't re-use the SaveMode in the group editor anymore.
As we only ever use SaveMode.CLOSE, I decided to remove this entirely.
We don't support re-loading at the moment, so it seems wrong to have
code in there that pretends to be able to do it.

Also I cleaned up some Eclipse warnings, so that the side-bar has fewer
yellow lines

Change-Id: I263b716c2edc98682bd5f5d1bc2b18ede15d7ced