Fix call-log auto-scroll-to-top behavior.
The auto-scroll-to-top behavior should only occur when displaying the
call-log immediately after a call finishes.  It was happening too
often, but is now fixed.
Bug: 5421613
Bug: 4088575

Workaround for bug whereby the smooth-scroll to top is broken when
invoked immediately after setting the list position (as is done to set
the position to "5" so that the smooth-scroll doesn't traverse the
entire list).
Bug: 6086980

Change-Id: I8dcc6f2cf84935875697b2454ba602bed03b66c4