Disable voicemail related receiver on tablet.
If the device is not a phone, we should not listen for the NEW_VOICMAIL
and the BOOT_COMPLETED broadcasts since we should not show or update the
notifications on a non-phone device.

This is because the call log is not shown on the device, so there is no
way to access the voicemails anyway.

Bug: 6044695
Change-Id: I606ea0646cea350f02c18d48d5d1eb43033b1b89