Prevent ProgressDialog from causing crash.
This is a minimal fix (or a workaround) for the bug 3129933:
SpecialCharSequenceMgr can crash Contacts app with a bad
Dialog.dismiss() call

Ideally we should merge two implementations into one and introduce
better cleanup mechanism for the whole class. Maybe we want to
instanciate the whole SpecialCharSequenceMgr, while it will cause
a lot of work. Note that this crash won't happen on Phone side,
so this kind of cleanup won't be needed there.

To keep the two implementations basically same, this change just has
*static* cleanup() method in Contacts part of the class.

Bug: 3129933
Change-Id: I99e4e1494f91f58f10f6ebbf814d2c8695e021f6