Fix jank when expanding photo in ContactDetailActivity.
The bug is caused as follows.  Tapping the photo invokes a
PhotoSelectionActivity.  When it initially sets the photo, a relayout
request is made, which causes a jank if the animation is already
underway.

Instead, we now wait for the layout before starting the animation.

Add new utility class with static doAfterLayout() method, since this
code was duplicated in a couple of places.

Bug: 6000382
Change-Id: Ic32a3db1f8c0be99fe0c160debede71f882c73e7