Allow setting photo on otherwise-empty local contacts.
Fix bug in ContactEditorFragment where, after tapping "Done",
the contact's photo is still unset.  This regression is due to
the changes to support high-res photos... now that we use a
side-channel to save the photo, the EntityDeltaList doesn't
know about the change.  The fix is to also create a thumbnail
and add it to the delta-list.  This allows us to avoid ugly
special-case code in the editor and save-service.  The high-res
photo is still saved via the same side-channel.

Bug: 6109553
Change-Id: I59bb9885eeb0260861ee167d3b5f1c7319896488