Relax phone number collapser
The current collapser basically compares two phone numbers for equality
using their normalzied numbers, but this can cause an issue when you have a
readonly contact from a different country who has a local version of their
phone numer in their profile.  Even if you have an international version
of their phone number, it may be overwritten on the contacts by the local
vesion.

So let's relax it.

The new collapser merges two numbers only if the "unformatted"
(i.e. ignoring all formatting characters) version of two phone numbers
are the same.

The following numbers will no longer be merged.  (All assume the user
is in the US.)
- +81-45-381-0000 and 011-81-45-381-0000
- 1-555-123-4567 and 555-123-4567
- +1-555-123-4567 and 1-555-123-4567

But the following will.
- 555-123-4567 and 5551234567

This affects:
- Contact card
- Phone number disambiguation on the phone app
- Quick contact

The list view on the phone app won't be affected, as it's done in SQL, group-by.

Bug 5779336

Change-Id: I2d1f2fcf8a8d85207383cb05a86f66c17af8acfe