Fix crash in StreamItemEntry.checkDecoded()
This method is meant to check if decodeHtml() was called prior to
getDecodedText() / getDecodedComments() calls.

The logic was flawed because HtmlUtils.fromHtml will return null
for empty strings, so even when decodeHtml() was called and mComments
wasn't null, mDecodedComments may still be null.

Let's just not try to be smart and introduce a new flag for this purpose.

Also changed the test-only constructor to a static method to make sure
it won't get called from the main apk.

Bug 6367860

Change-Id: I9a774e010666bd5a83ef916e817b50052afd325f