Added menu option to clear frequents in favorites tab of Phone and People
The clear frequents menu option is only visible when there are items in
the frequently contacted list.  When the user selects the menu option, a
confirmation dialog is shown to verify whether to actually clear frequents.

New ClearFrequentsDialog class created in a new com.android.contacts.dialog
package that is shared by both Phone and People apps to show the
confirmation dialog of whether to clear frequents.

Bug:5475575
Change-Id: I876d3fe15f16635570f3ecb7cd8ed2a19227e316