Fixed bug when shortcut to contact has no display name
This happens when there is no displayed name for a contact (e.g., when the
contact only has a physical address but no phone number or email).  This
caused a null pointer exception to be thrown when Launcher2 receives the
broadcasted intent to create a shortcut with a null display name.

If the display name is found to be empty (i.e., null or zero-length),
the string "(No name)" (from R.string.missing_name) is considered to be
the display name, which is consistent with what the People app shows as
the display name.

Bug:6346747
Change-Id: I82c9782cf88d630bd232be323b961809214da8d2