Support adding phone/email to readonly contacts
If the selected contact has no editable raw_contacts, create one on the
first editable account found, and add the data to it.

We'll reuse the "contact is readonly" mode if there's no editable accounts
on the system.  We could improve the message, but if there's no editable
accounts we shouldn't show the contact picker in the first place, so I
didn't change the message.  Let's fix the picker instead.

Tested:
(Previously supported casses)
- Add a phone number to an existing editable contact.
- Add a phone number to an existing local-only contact.

(Newly supported cases)
- Add a phone number to an existing read-only contact.
- Try to add a phone number to an existing read-only contact, but no
editable accounts are set up.

Bug 5462347

Change-Id: I55db01610cebb60d3830e8ec4ac6bf1a3dfdad88