Added help menu options to People app
This has been implemented with the idea to enable the creation of help
menu options in a standard way throughout the Android platform.  One
particularly important feature is to hide the exact URL that is being
used to invoke the help pages via the browser via using vendor-specific
overlays of string resources.

Also, other standard context for the help like the preferred language and
app version code are incorporated automatically via the HelpUtils function.

Bug:6206323
Change-Id: I5a1fff4d99ac93d3a6aa05da91f465fad072cb8a