Made layout fixes to support 7" tablets
Renamed resource directory suffixes for 7" tablets in landscape
orientation from sw580dp-w1000dp to sw580dp-w940dp.

In Group and All panes, the ratio of the left list pane to right detail
pane is 1:2.  Set up a threshold of the minimum width to show images in
the browse list under the All tab and show/hide images appropriately.

Show 3 columns of images in the Group detail pane when the 7" tablet is
in landscape orientation.

Width in left scrollbar for left pane browse list reduced for 7" tablet.

Bug: 6398940
Bug: 6378492
Change-Id: I7730830470e794a59cf7b2b910ae7aadf82b36c6