Hide scrollbar in search mode when no contacts in Phone
This hides the scrollbar when there are no contacts in search mode.

Some refactoring has been done to make it easier to tell what menus
are made visible for each of the Phone tabs and when in search mode.
Intents and onClick listeners have been moved to the
onCreateOptionsMenu since these do not change.  Also, corrected the
test for isEmpty() in PhoneFavoriteMergedAdapter which allows the
message "No contacts" to show up when there are no contacts to list
under the Phone favorites tab.

Bug: 6036160
Change-Id: I43bbd1bd56b009736cc2a966a3ffb0ad6630792c