Close photo selection activity on config change
This activity doesn't handle orientation changes right.  Let's just close it
on any configuration changes.   This activity is kinda like a popup, so it
should be okay.

Now this activity handles all config changes by itself, so we can remove a lot
of things, such as onSaveInstanceState and PendingPhotoResult.

Bug 6236774

Change-Id: Ib7037a66d092c8d9a285224ea4cf70e7da8ff1b5