Prevent action-bar icons from changing while QuickContacts active.
We were setting a flag in onPause() to ensure that we re-initialize the UI
properly in onResume().  However, this was causing a side-effect when the
background PeopleActivity received loader-notifications that were triggered
by the foreground QuickContactsActivity, because the code triggered by the
loader-notification was looking at the flag set in onPause().

To fix this, showContactsUnavailableFragmentIfNecessary(), gains a boolean
"forceUpdate" argument, which more directly reflects the intended usage.

Also, rename sCUFIN() to updateViewConfiguration().  It's cleaner.  And more
accurate.

Bug: 6117162
Change-Id: I35b9f06c0deeffe94d5cf80d4df58a9866387792