Fixed group editing popup from contact editor
Removed checkbox next to the last "Create new group" option
in popup and removed square brackets in string.

After creating a new group, the view shows the popup again
with the newly create group visible and checked.  This is
accomplished via a listener interface to the
GroupCreationDialogFragment.  The tricky part to this is if
the user changes device orientation while the
GroupCreationDialogFragment is open.  The ContactEditorFragment
will save the resource id of the RawContactEditorView (since
there can be many via joins) and the view id of the
GroupMembershipView which is the listener.

Bug: 6013015
Change-Id: Ia50a68fb6f4631995ae913182765efb24ef227ec