Task jumping in contact details
When viewing contact details from a quick contact card, the contact
details appear in a two-pane UI, forming a sort of implicit Up
navigation. As a result, viewing contact details from another task is
in effect a task jump since the user may choose to navigate to other
contacts using the other pane of the UI.

Use the new shouldUpRecreateTask API in JB to determine whether the
trampoline from ContactDetailActivity to the two-pane People activity
should also include a task jump.

Bug:6449430

Change-Id: I90b86873a3a21022d52bb387d9adc5391c8d23ac